ai-service/requirements.txt
tensorflow==2.21.0
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.19
pillow==11.0.0
numpy==2.4.2
matplotlib==3.10.0
scikit-learn==1.6.0